[DeploySummary] Inline all code, instead of relative imports across rule files
